webapi/tct-svg-html5-tests/svg/w3c/foreign_content_010.html

<!DOCTYPE html>
<html>
    <head>
        <title>HTML 5 Foreign Content SVG in HTML </title>
        <meta description="Test to verify SVG Self Closing tags parses properly" />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
        <script type="text/javascript">
            var t = async_test(document.title, {timeout: 2000});
            function RunTest()
            {
                try
                {
                    var parentNode = document.getElementById("svg1");
                    t.step(function (){
                        assert_true(parentNode.childNodes[1].localName=="rect" && parentNode.childNodes[3].localName=="circle", "Test to verify SVG Self Closing tags parses properly");
                    }, document.title);
                    t.done();
                }
                catch(ex)
                {
                    t.step(function (){
                        assert_true(false, "there is a exception");
                    }, document.title);
                    t.done();
                }
            }
        </script>
    </head>

    <body onload="RunTest()">
        <div id="log"></div>
        <svg id="svg1" width="100px" height="100px">
            <rect id="rect1" />
            <circle id="circle1" />
        </svg>

    </body>
</html>
